Port PoolAllocator
There are very obvious differences in how the ported PoolAllocator
functions in the port. Instead of passing in a size, we pass in a
Bump Type, from which it can then create items as needed.

We are also not preallocating the space, but rather create objects as
needed, and storing free'd objects for reuse later.