Remove allocations from Dbvt rayTest
Recycle Vector3 and array objects. The `stack` array is handled as
a special case with its own memory pool. This is because the stack
array needs to be resized and could end up being large. By
recycling it and only resizing when we need the length to increase,
we can make it so that the resize only happens a finite number of
times, until there exist arrays of sufficient length for all
future ray tests.